Implement Spring integration tests
