+++ java-cfenv-boot/src/main/resources/META-INF/spring.factories
@@ -6,3 +6,6 @@ org.springframework.boot.env.EnvironmentPostProcessor=\
 org.springframework.context.ApplicationListener=\
   io.pivotal.cfenv.spring.boot.CfDataSourceEnvironmentPostProcessor,\
   io.pivotal.cfenv.spring.boot.CfEnvironmentPostProcessor
+
+io.pivotal.cfenv.spring.boot.CfEnvProcessor=\
+  io.pivotal.cfenv.spring.boot.RedisCfEnvProcessor
